+++ src/natural/GA/crossover/CrossoverInterface.java
@@ -4,6 +4,6 @@ import natural.GA.Individual;
 
 @SuppressWarnings("unused")
 public interface CrossoverInterface {
-    void crossover(int[] preCalc, Individual male, Individual female, Individual baby);
+    void crossover(long[] preCalc, Individual male, Individual female, Individual baby);
 }
