+++ src/natural/GA/preCalc/PreCalcs.java
@@ -8,10 +8,10 @@ public class PreCalcs {
 
     public static PreCalcInterface minAndSum(){
         return (individuals) -> {
-            int[] result = new int[2];
-            result[0] = Arrays.stream(individuals).mapToInt(AbstractIndividual::getFitness).min().orElse(0);
-            int absMin = Math.abs(Math.min(0, result[0]));
-            result[1] = Arrays.stream(individuals).mapToInt(i -> i.getFitness() + absMin).sum();
+            long[] result = new long[2];
+            result[0] = Arrays.stream(individuals).mapToLong(AbstractIndividual::getFitness).min().orElse(0);
+            long absMin = Math.abs(Math.min(0, result[0]));
+            result[1] = Arrays.stream(individuals).mapToLong(i -> i.getFitness() + absMin).sum();
             return result;
         };
     }
